requirements.txt
alembic==1.4.2
cffi==1.14.2
click==7.1.2
colorama==0.4.5
dateUts==0.2.8
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.8.1
